+++ src/pretalx/event/models/event.py
@@ -249,6 +249,7 @@ class Event(LogMixin, models.Model):
         frab_json = '{export}schedule.json'
         frab_xcal = '{export}schedule.xcal'
         ical = '{export}schedule.ics'
+        widget_data_source = '{schedule}widget/v1.json'
 
     class orga_urls(EventUrls):
         create = '/orga/event/new'